penalty done.. voice disconnect needed

+++ commands/clear.js
@@ -1,3 +1,4 @@
+const validation = require("../validation/validator");
 const errorHandler = require("../handlers/error_handler");
 module.exports = {
   name: "clear",
@@ -5,19 +6,24 @@ module.exports = {
   async execute(bot, message, args) {
     // console.log(args);
 
-    if (!args[0]) args[0] = 5;
-    if (isNaN(args[0])) return message.reply("please enter number");
-    if (args[0] > 100) return message.reply("please enter lower than 100");
-    if (args[0] < 1) return message.reply("please enter greater than 0");
+    if (validation.admin(message)) {
+      if (!args[0]) args[0] = 5;
+      if (isNaN(args[0])) return message.reply("please enter number");
+      if (args[0] > 100) return message.reply("please enter lower than 100");
+      if (args[0] < 1) return message.reply("please enter greater than 0");
 
-    try {
-      await message.channel.messages
-        .fetch({ limit: args[0] })
-        .then((messages) => {
-          message.channel.bulkDelete(messages);
-        });
-    } catch (error) {
-      errorHandler(message, error);
-    }
+      try {
+        await message.channel.messages
+          .fetch({ limit: args[0] })
+          .then((messages) => {
+            message.channel.bulkDelete(messages);
+          });
+      } catch (error) {
+        errorHandler(message, error);
+      }
+    } else
+      message.channel.send(
+        `<@${message.author.id}> You don't have permission.`
+      );
   },
 };
