+++ commands/server_config.js
@@ -15,7 +15,7 @@ module.exports = {
           let data = await serverData.findOne({ id: server_id });
 
           if (!data) {
-            createChanneles(message);
+            await createChanneles(message);
           } else {
             message.channel.send("This server already added!");
           }
@@ -53,6 +53,8 @@ module.exports = {
 
           roles: {
             common_lobby: obj.view_waiting,
+            bot_admin: obj.bot_admin,
+            view_bot_category: obj.view_bot
           },
 
           penalty_id: [],
@@ -74,69 +76,83 @@ module.exports = {
           message.guild.roles
             .create({
               data: {
-                name: "view_bot_category",
-                color: "#00d9fc",
+                name: "bot_admin",
+                color: "#ff6600",
               },
             })
-            .then((view_bot_role) => {
-              id_obj.view_bot = view_bot_role.id;
+            .then((bot_admin) => {
+              id_obj.bot_admin = bot_admin.id;
             })
             .then(() => {
-              message.guild.channels
-                .create("Warzone BOT", {
-                  type: "category",
-                  permissionOverwrites: [
-                    {
-                      id: everyoneRole.id,
-                      deny: ["VIEW_CHANNEL"],
-                    },
-                    {
-                      id: id_obj.view_bot,
-                      allow: ["VIEW_CHANNEL"],
-                    },
-                  ],
+              message.guild.roles
+                .create({
+                  data: {
+                    name: "view_bot_category",
+                    color: "#00d9fc",
+                  },
                 })
-                .then((category) => {
+                .then((view_bot_role) => {
+                  id_obj.view_bot = view_bot_role.id;
+                })
+                .then(() => {
+                  // console.log(everyoneRole.id)
+                  // console.log(id_obj.view_bot);
                   message.guild.channels
-                    .create("commands", {
-                      type: "text",
+                    .create("Warzone BOT", {
+                      type: "category",
+                      permissionOverwrites: [
+                        {
+                          id: everyoneRole.id,
+                          deny: ["VIEW_CHANNEL"],
+                        },
+                        {
+                          id: id_obj.view_bot,
+                          allow: ["VIEW_CHANNEL"],
+                        },
+                      ],
                     })
-                    .then((ch_1) => {
-                      ch_1.setParent(category.id);
-                      id_obj.commands = ch_1.id;
-                      id_obj.parent = category.id;
-                    })
-                    .then(() => {
+                    .then((category) => {
                       message.guild.channels
-                        .create("team_mode_selection", {
+                        .create("commands", {
                           type: "text",
                         })
-                        .then((ch_2) => {
-                          ch_2.setParent(category.id);
-                          id_obj.team_mode_selection = ch_2.id;
+                        .then((ch_1) => {
+                          ch_1.setParent(category.id);
+                          id_obj.commands = ch_1.id;
+                          id_obj.parent = category.id;
                         })
                         .then(() => {
                           message.guild.channels
-                            .create("waiting_lobby", {
-                              type: "voice",
+                            .create("team_mode_selection", {
+                              type: "text",
                             })
-                            .then((ch_3) => {
-                              ch_3.setParent(category.id);
-                              id_obj.waiting_lobby = ch_3.id;
+                            .then((ch_2) => {
+                              ch_2.setParent(category.id);
+                              id_obj.team_mode_selection = ch_2.id;
                             })
                             .then(() => {
-                              message.guild.roles
-                                .create({
-                                  data: {
-                                    name: "view_waiting",
-                                    color: "#00d9fc",
-                                  },
+                              message.guild.channels
+                                .create("waiting_lobby", {
+                                  type: "voice",
                                 })
-                                .then((role_1) => {
-                                  id_obj.view_waiting = role_1.id;
+                                .then((ch_3) => {
+                                  ch_3.setParent(category.id);
+                                  id_obj.waiting_lobby = ch_3.id;
                                 })
-                                .then(async () => {
-                                  await saveData(id_obj);
+                                .then(() => {
+                                  message.guild.roles
+                                    .create({
+                                      data: {
+                                        name: "view_waiting",
+                                        color: "#00d9fc",
+                                      },
+                                    })
+                                    .then((role_1) => {
+                                      id_obj.view_waiting = role_1.id;
+                                    })
+                                    .then(async () => {
+                                      await saveData(id_obj);
+                                    });
                                 });
                             });
                         });
@@ -163,7 +179,11 @@ module.exports = {
                     (ch) => (ch.id = data.channels[key])
                   );
                   // console.log(data.channels[key]);
-                  channel.delete();
+                  if (channel) {
+                    channel.delete();
+                  } else {
+                    console.log(channel);
+                  }
                 }
               }
             }
@@ -175,7 +195,9 @@ module.exports = {
                     (ch) => (ch.id = data.roles[key])
                   );
                   // console.log(data.channels[key]);
-                  role.delete();
+                  if (role) {
+                    role.delete();
+                  } else console.log(role);
                 }
               }
             }
